+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.6.1
+
+* Added human modelling classes SWKHmiWorkbench and SWKHumanCatalog. Please note these are not documented in the CATIA
+  V5 Basic Help file. Example of usage also added.
+* Added new class method Application.input_box(). @ptm-tm
+* fixed test issue with cat materials where catia env has multiple start-up paths.
+* fixed issue with HybridShapeFactory.add_new_sphere for cases where an axis isn't defined.
+* fixes to RenderingMaterials class methods. @deloarts
+* added new enum type geometrical_feature_type to support HybridShapeFactory.get_geometrical_feature_type().
+
 ## 0.6.0
 
 * Removed deprecated method Product.get_products() use Product.products instead.